make skip_cond_incl work
Normally, read_expr() regard '\n' as a whitespace and goes over it to the next token in the next line.
But when evaluating #if condition, it leads the parser's violation.

So this commit prepares a special buffer named "altbuffer" to store the const expression before '\n'.
In order to do that, read_token_int should have a mode named "return_at_eol".